troubleshoot 32bit builds

+++ pandas/tests/frame/indexing/test_setitem.py
@@ -935,8 +935,8 @@ class TestDataFrameSetitemCopyViewSemantics:
         df_view = df[:]
         df[["B"]] = value
 
-        expected = DataFrame([[0, 1], [0, 1]], columns=cols, dtype=np.intp)
-        assert (expected.dtypes == np.intp).all()  # troubleshoot 32bit builds
+        expected = DataFrame([[0, 1], [0, 1]], columns=cols, dtype=int)
+        assert (expected.dtypes == int).all()  # troubleshoot 32bit builds
         tm.assert_frame_equal(df, expected)
         tm.assert_frame_equal(df_view, df_copy)
